+++ i18n/en.pot
@@ -5,13 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-<<<<<<< HEAD
 "POT-Creation-Date: 2024-11-13T08:29:01.348Z\n"
 "PO-Revision-Date: 2024-11-13T08:29:01.348Z\n"
-=======
-"POT-Creation-Date: 2024-11-06T18:43:30.317Z\n"
-"PO-Revision-Date: 2024-11-06T18:43:30.318Z\n"
->>>>>>> f9bc738 (feat(dataSet): dataSet sectionedForm)
 
 msgid "schemas"
 msgstr "schemas"
@@ -424,6 +419,15 @@ msgstr "Translation updated successfully"
 
 msgid "Save translations"
 msgstr "Save translations"
+
+msgid "Go back"
+msgstr "Go back"
+
+msgid "Next section"
+msgstr "Next section"
+
+msgid "Save and exit"
+msgstr "Save and exit"
 
 msgid "Can edit and capture"
 msgstr "Can edit and capture"
@@ -1177,6 +1181,12 @@ msgstr "Configure data elements"
 msgid "Choose what data is collected for this data set."
 msgstr "Choose what data is collected for this data set."
 
+msgid "Configure data entry periods"
+msgstr "Configure data entry periods"
+
+msgid "Choose for what time periods data can be entered for this data set"
+msgstr "Choose for what time periods data can be entered for this data set"
+
 msgid "Setup"
 msgstr "Setup"
 
@@ -1188,6 +1198,9 @@ msgstr "Data Elements"
 
 msgid "Periods"
 msgstr "Periods"
+
+msgid "Period type"
+msgstr "Period type"
 
 msgid "Organisation Units"
 msgstr "Organisation Units"
