+++ src/lib/form/index.ts
@@ -6,4 +6,3 @@ export { validate, createFormValidate } from './validate'
 export { useOnSubmitEdit, useOnSubmitNew } from './useOnSubmit'
 export { modelFormSchemas } from './modelFormSchemas'
 export * from './sectionedForm'
-export * from './sectionedForm/sectionedFormDescriptor'
